Move the Jitm UTM provider to the jitm package
