+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/jitm/JitmUtmProvider.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.mystore
+package com.woocommerce.android.ui.jitm
 
 import com.woocommerce.android.util.UtmProvider
 import dagger.Reusable
@@ -8,7 +8,7 @@ import javax.inject.Inject
  * This is just a wrapper around the UtmProvider to allow mocking the implementation in tests
  */
 @Reusable
-class MyStoreUtmProvider @Inject constructor() {
+class JitmUtmProvider @Inject constructor() {
     fun getUrlWithUtmParams(
         source: String,
         id: String,
